Safari-609.1.11.1/LayoutTests/fast/dom/Attr/parent-adopt-node-expected.txt

PASS attr.ownerDocument.title is anotherDocument.title
PASS successfullyParsed is true

TEST COMPLETE
